+++ pkg/providers/property/testdata/TestDataPropertyIncludes/no_group_id.tf
@@ -1,0 +1,7 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+data "akamai_property_includes" "test" {
+  contract_id = "contract_123"
+}