+++ pkg/providers/property/testdata/TestDataPropertyIncludes/with_parent_property/list_available_includes_type_microservices.tf
@@ -1,0 +1,13 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+data "akamai_property_includes" "test" {
+  contract_id = "contract_123"
+  group_id    = "group_321"
+  parent_property {
+    id      = "property_id_123"
+    version = 10
+  }
+  type = "MICROSERVICES"
+}