+++ pkg/providers/property/testdata/TestDataPropertyIncludes/without_parent_property/list_includes_type_common_settings.tf
@@ -1,0 +1,9 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+data "akamai_property_includes" "test" {
+  contract_id = "contract_123"
+  group_id    = "group_321"
+  type        = "COMMON_SETTINGS"
+}